Update pushButtonCounter.ino
The new debounce mechanism functions similarly to a simple delay(), but the program doesn't pause, allowing the Arduino to continuously update the display and poll button readings.

Previously, direction checking logic was outside of the button press conditional; this led to Arduino constantly changing direction of counting at 0 and 9 even when the button was not being pressed. Additionally, checking that the count direction is wrong before changing it saves the program from incorrectly changing direction if the debounce mechanism fails.